+++ changelog.md
@@ -5,6 +5,7 @@ Last change: 18-MAY-2026 CEST
 ## 0.10.1
 
 - SPINNA: comparing models uses fitting modes and has cleaner progress dialog
+- SPINNA: convenient fitting of LE
 
 ## 0.10.0